# apispecs/neu-connect-api-docs.yaml
openapi: 3.1.0
info:
  version: 0.0.1
  title: NEU Connect API
servers:
  - url: http://localhost:8080
paths:
  /api/posts:
    post:
      summary: Create a new Post
      description: Create a new Post
      tags:  # Associate this operation with the "Post endpoints" tag
        - Post API
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Post"
      responses:
        201:
          description: Post created successfully
        401:
          description: Unauthorized
        403:
          description: Forbidden
  /api/posts/{id}:
    get:
      summary: Get a specific Post
      description: Get a specific Post
      tags:  # Associate this operation with the "Post endpoints" tag
        - Post API
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the Post to retrieve
          schema:
            type: integer  # Assuming the ID is of type integer
      responses:
        200:
          description: Post fetched successfully
        401:
          description: Unauthorized
        403:
          description: Forbidden
...
  /api/posts/{id}/comment:
    post:
      summary: Add Comment to Post
      description: Add Comment to Post
      tags:  # Associate this operation with the "Post endpoints" tag
        - Post API
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the Post to retrieve
          schema:
            type: integer  # Assuming the ID is of type integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Comment"
      responses:
        200:
          description: Post fetched successfully
        401:
          description: Unauthorized
        403:
          description: Forbidden
...
  /api/products:
    post:
      summary: Create a Product
      description: Create a Product
      tags:  # Associate this operation with the "Post endpoints" tag
        - Product API
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
      responses:
        201:
          description: Created a Product successfully
        401:
          description: Unauthorized
        403:
          description: Forbidden
  /api/products/{id}:
    get:
      summary: Get a specific Product
      description: Get a specific Product
      tags:  # Associate this operation with the "Post endpoints" tag
        - Product API
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the Product to retrieve
          schema:
            type: integer  # Assuming the ID is of type integer
      responses:
        200:
          description: Post fetched successfully
        401:
          description: Unauthorized
        403:
          description: Forbidden
  /api/products/fetch:
    get:
      summary: Get paginated List of Product
      description: Get paginated List of Product
      tags:  # Associate this operation with the "Post endpoints" tag
        - Product API
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaginationOptions'
      responses:
        200:
          description: Product fetched successfully
        401:
          description: Unauthorized
        403:
          description: Forbidden
...
components:
  schemas:
    Post:
      type: object
      required:
        - title
        - content
      properties:
        title:
          type: string
        content:
          type: string
        upvote:
          type: integer
          default: 0
        comments:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
        author:
          type: string  # Assuming the User ID is represented as a string
        createdAt:
          type: string  # Assuming the date is represented as a string
          format: date-time
        updatedAt:
          type: string  # Assuming the date is represented as a string
          format: date-time
        tag:
          type: array
          items:
            type: string
    Comment:
      type: object
      required:
        - comment
      properties:
        comment:
          type: string
        author:
          type: string  # Assuming the User ID is represented as a string
        likes:
          type: integer  # Likes is int type
        parentComment:
          $ref: '#/components/schemas/Comment'
    Product:
      type: object
      properties:
        karma:
          type: integer
        title:
          type: string
        description:
          type: string
        images:
          type: array
        createdAt:
          type: string  # Assuming the date is represented as a string
          format: date-time
        updatedAt:
          type: string  # Assuming the date is represented as a string
          format: date-time
        deletedAt:
          type: string  # Assuming the date is represented as a string
          format: date-time
...
    PaginationOptions:
      type: object
      properties:
        pageSize:
          type: integer
        pageNumber:
          type: integer
...